requirements.txt
Flask==1.1.2
gunicorn==20.0.4
google-api-python-client==1.12.8
urllib3==1.25.10
google-cloud-bigquery==1.27.2
google-cloud-storage==1.30.0
python-dateutil==2.8.1
gdown=3.12.2
